Time is money!
This will try, rather unelegant if I may say so, to avoid all the really annoying tasks. It will skip the main menu, skip the "Click me to place stuff" and last but not least it will skip the last part, where you just have to wait and acknowledge something. I can't even recall what it was!